fixed usage text in nibconv
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@200 048eaad1-c872-4173-8c30-523bb29f53ec

+++ prot.c
@@ -17,7 +17,7 @@ BYTE density_map_rapidlok[MAX_TRACKS_1541] = {
 	2, 2, 2, 2, 2, 2, 2				/* 36 - 42  */
 };
 
-int skew_map[MAX_TRACKS_1541] = {
+BYTE skew_map[MAX_TRACKS_1541] = {
 	0, 0, 0, 0, 0, 0, 0, 0, 0, 0,  /*  1 - 10 */
 	0, 0, 0, 0, 0, 0, 0, 0, 0, 0,	/* 11 - 20 */
 	0, 0, 0, 0, 0, 0, 0, 0, 0, 0,	/* 21 - 30 */
